Fix lint errors aside from non-null assertions
